# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspace_id}:
    get:
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
    patch:
      operationId: updateWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: Updated workspace profile
          content:
            application/json:
              schema:
                type: object
                required: [workspace]
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
                  event:
                    type: object
        "403":
          description: Workspace manager permission required
    delete:
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "204":
          description: Workspace metadata deleted; queued upload-object cleanup may continue asynchronously
        "403":
          description: Workspace owner permission required
...
  /api/channels/{channel_id}/pull-request-status:
    get:
      operationId: getPullRequestStatus
      parameters:
        - $ref: "#/components/parameters/channel_id"
      responses:
        "200":
          description: Live status for the linked GitHub pull request
          content:
            application/json:
              schema:
                type: object
                required: [pull_request]
                properties:
                  pull_request:
                    $ref: "#/components/schemas/PullRequestStatus"
        "400":
          description: Channel has no valid linked GitHub pull request
        "401":
          description: Authentication required
        "403":
          description: Channel read access required
        "502":
          description: GitHub status is temporarily unavailable
...
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    channel_id:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    PullRequestStatus:
      type: object
      required: [state, ci_state, checks_total, review_state, last_reply_author, last_reply_at, updated_at]
      properties:
        state:
          type: string
          enum: [draft, open, closed, merged]
        ci_state:
          type: string
          enum: [passing, failing, pending, not_reported, unknown]
        checks_total:
          type: integer
          minimum: 0
        review_state:
          type: string
          enum: [pending, approved, changes_requested]
        last_reply_author:
          type: string
        last_reply_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateWorkspaceRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
...
    Workspace:
      type: object
      required: [id, route_id, name, slug, icon_url, created_at]
      properties:
        id:
          type: string
        route_id:
          type: string
          description: Immutable public route ID used in app URLs.
        name:
          type: string
        slug:
          type: string
        icon_url:
          type: string
        created_at:
          type: string
          format: date-time
        role:
          type: string
          enum: [owner, moderator, member, guest, bot]
...